runtime: rename modules, cleaner namespace
